requirements.txt
numpy
pandas
matplotlib
tqdm
imutils
tensorflow
keras
opencv-python
Pillow
scikit-learn
pickle5
streamlit
seaborn
plotly
shap
lime
streamlit